build: configure flavors for all modules
